Upgrade debugger visualizer
* Split debugger visualizer into the debugger side (targeting .net 4.7.2) and debuggee side (targeting .net standard 2.0).
* Remove NRules references from the debugger-side visualizer, and just pass it a serialized rete graph as an XML string.
* Move DGML writer to NRules assembly.
* Reference Visual Studio assemblies from under the VS installation
