+++ src/NRules.Debugger.Visualizer/NRules.Debugger.Visualizer.DebuggeeSide/Properties/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("NRules.Debugger.Visualizer")]
+[assembly: AssemblyTitle("NRules.Debugger.Visualizer.DebuggeeSide")]
 [assembly: AssemblyDescription("Visual Studio Debugger Visualizer for NRules")]
 [assembly: ComVisible(false)]
