--- src/NRules.Debugger.Visualizer/NRules.Debugger.Visualizer/SessionObjectSource.cs
@@ -1,16 +1,0 @@
-﻿using System.IO;
-using Microsoft.VisualStudio.DebuggerVisualizers;
-using NRules.Diagnostics;
-
-namespace NRules.Debugger.Visualizer
-{
-    public class SessionObjectSource : VisualizerObjectSource
-    {
-        public override void GetData(object target, Stream outgoingData)
-        {
-            var session = (ISessionSnapshotProvider) target;
-            var snapshot = session.GetSnapshot();
-            base.GetData(snapshot, outgoingData);
-        }
-    }
-}
